add view file and update core
